test: move schema inspect live checks to integration
